# encode performance tweaks