**update:  merged** ~~Depends on #44, so review that first.~~

Inspired by the work in #40, which had several different changes in it, with mixed results. 

I've tried to extract only the strictly beneficial parts here.

```
$ cargo bench --bench="*" -- --baseline=main3
    Finished bench [optimized] target(s) in 0.03s
     Running benches/benchmarks.rs (target/release/deps/benchmarks-594cd3d509514e72)
encode 10_000 coordinates at precision 1e-5
                        time:   [108.92 µs 109.15 µs 109.42 µs]
                        change: [-14.204% -14.035% -13.852%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

encode 10_000 coordinates at precision 1e-6
                        time:   [124.45 µs 125.23 µs 126.08 µs]
                        change: [-8.1045% -7.5507% -6.9785%] (p = 0.00 < 0.05)
                        Performance has improved.

#### NOTE: This is likely spurious — I only made changes to the `encode` code path.
decode 10_000 coordinates at precision 1e-5
                        time:   [89.408 µs 90.392 µs 91.407 µs]
                        change: [-4.8269% -3.8252% -2.8511%] (p = 0.00 < 0.05)
                        Performance has improved.

decode 10_000 coordinates at precision 1e-6
                        time:   [93.003 µs 93.917 µs 94.787 µs]
                        change: [-1.4427% -0.5576% +0.2285%] (p = 0.20 > 0.05)
                        No change in performance detected.
```